+++ CHANGELOG.md
@@ -4,6 +4,10 @@
  - Search Result Page, Category Page
   - add Products per Page configurartion which allows user to define custom configuration without any change in code
 
+### Fix
+ - Export
+  - fix `Model\Export\Catalog\ProductType\ConfigurableDataProvider::getChildren` throws an SQL syntax error if configurable product has no variants assigned   
+
 ## [v2.5.0] - 2022.03.04
  - Import 
   - fix error while push import feed using CLI after generate it for API version 7.x